moved create postal code auto lookup to view model
